improve attributes ergonomics
attributes are pretty arbitrary and parsing depends on the netlink
message, so we cannot have one type for attributes. Instead, we provide
an Attribute trait, and implement parsing/emitting for types that
implement this trait. In addition, we provide a default type for
attributes, so is just a buffer of bytes.

To make iterating of attributes efficient and pleasant, add an iterator
that does not actually perform any parsing.