remove bound checks when accessing attribute value
Parsing and emitting attributes is the bread and butter of netlink.

This change makes emitting faster but susceptible to crash if the
emitted packet is malformed. I think it is fine, because anyway, we
should not have attribute types that creates malformed packets.

Note that DefaultAttribute::from_packet and Repr::parse still implements
bound checks though, because we cannot trust input packets and must not
crash upon receiving them.