get rid of LinkData
I thought it would be nice to have a type Link that holds a netlink
connection and provides some methods like Link::modify(...):

impl Link {
  fn modify(&mut self) -> impl Future<Item=(), Error=NetlinkIpError> {
    // ...
  }
}

However, the future we return would need to hold a mutable reference to
the link, so that when it resolves the link is updated. Something like:

struct LinkFuture<'a> {
  link: &mut Link
}

impl<'a> Future for LinkFuture<'a> {
  type Item = ();
  type Error = NetlinkIpError;
  // ...
}

All this might be possible but since Tokio requires futures to have a
'static lifetime, we could not spawn such a future. Since our future
cannot hold a reference to the link, it must hold the link itself, and
return it when it finished:

impl Link {
  fn modify(self) -> impl Future<Item=Link, Error=NetlinkIpError> {
    // ...
  }
}

But that makes the API much less user-friendly:

- there would be no way to start multiple requests on the link
- error handling is tricky: if an error occurs, we basically lost the
  link unless we cloned it before making the request. To prevent that,
  our error would need to contain the link but that get a little
  weird...

So instead of this kind of API, we falling back to a more basic one: the
ConnectionHandle will have all the methods to send the actual requests.
That allows us to get rid of the current Link type, and rename LinkData
into Link.

To be future proof, we do not actually implement all the methods on
ConnectionHandle, because that would lead to too many methods (get_link,
get_routes, get_address, add_address, add_link, etc. etc.). Instead
we'll create a special handle for each type of message: links,
addresses, routes, etc.