Ignore unexpected bytes in Nla buffers
This commit changes is changing the logic of the NativeNla::from_bytes()
method to truncate or pad with 0 the input buffer to match the size of
the result struct.

The result is that fields which are not defined in the struct are ignored
and fields which are defined in the struct but missing in the buffer are
set to 0. In this way the parsing cannot fail.

Fix #3
Fix #8