LinkAddRequest: Allow adding macvlan on a link
Following adds support to create MACVLAN on a link.
This is equivalent to `ip link add NAME name link LINK type macvlan mode MACVLAN_MODE`.
But instead of specifying a link name (LINK), we specify a link index
and a mode.

Veth must be already up for this to work. Following behaviour is
expected cause its same for `vxlan` and `vlan`.

Signed-off-by: Aditya Rajan <arajan@redhat.com>