Fix padding in NetlinkCodec
The definition of the `NLMSG_NEXT` macro in the kernel makes it clear
that netlink messages are padded to a multiple of 4 bytes; this padding
is not part of the `nlmsg_len` field of the message header.

Luckily `nlmsg_len` is usually already a multiple of 4 (as the netlink
attributes are also padded to multiple of 4 bytes, and fixed-length
parts at the beginning of messages are usually aligned too), so the
padding is usually empty.

Anyway, fix this:

1. Make sure encoded messages is padded to multiple of 4 bytes

Technically the last message in a datagram doesn't need padding (i.e.
the kernel doesn't require it), but it's easier to be consistent.

2. Skip padding in decoder

Similar to the kernel we ignore missing padding for now, which would be
real bad if the buffer is received via "stream" instead of "datagram",
and might include partial datagrams (i.e. where the padding is read
later and not yet in the buffer).

Netlink sockets are datagram sockets though, so this is fine.

Note: NetlinkAuditCodec might need similar changes, but audit is quite
special.