rtnl/link: add match_name in addition to add_name_filter
Since kernel commit [a3d1289126e7b][1], it is possible to set the
IFLA_IFNAME attribute in a RTM_GETLINK message, allowing to get a link
by its name without dumping the whole link list. In large setups, this
speeds up massively the library, going from "painfully slow" to "really
quick" (in my application with thousands of links, a lookup by name
went from hundreds of ms to hundreds of µs).

This is based off the equivalent function in [libnl][2].

[1]: https://github.com/torvalds/linux/commit/a3d1289126e7b14307074b76bf1677015ea5036f
[2]: https://github.com/thom311/libnl/blob/master/lib/route/link.c#L1376