rtnetlink: add a way to set netlink flags in requests
There has been several PRs to add helper methods to set the flags in
the netlink header, so it seems that this is something we should
provide systematically.

This PR introduces three new types to manipulate netlink flags:

- `NewFlags`: flags used for "set" and "add" requests
- `DelFlags`: flags used to "del" requests
- `GetFlags:` flags used for "get" requests

Each request now has a `set_flags` method to force a certain set of
flags. This may interfere with other builder methods for certain
requests, or may simply render the request ineffective, so the
documentation always warns users.