rtnetlink: LinkDelRequest set correct flags
NLM_F_EXCL and NLM_F_CREATE are not valid flags for the delete request.
On kernel v5.19 there was a new delete flag NLM_F_BULK added which has
the same value as NLM_F_EXCL. This causes the kernel to respond with
EOPNOTSUPP when trying to remove an interface.

Fixes #273

Signed-off-by: Paul Holzinger <pholzing@redhat.com>